+++ vite-project/src/pages/ProductListPage.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProductListPage = () => {
+  return <div className="max-w-screen-xl"></div>;
+};
+
+export default ProductListPage;
